对开发者分支做了一系列修正

+++ src/main/java/com/qc/rc/controller/InterviewController.java
@@ -6,6 +6,7 @@ import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
 
+import javax.mail.search.IntegerComparisonTerm;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 
@@ -19,29 +20,29 @@ import org.springframework.web.servlet.ModelAndView;
 
 import com.github.pagehelper.PageInfo;
 import com.qc.rc.common.FormParameterUtil;
+import com.qc.rc.common.GetUser;
+import com.qc.rc.common.GetUuid;
 import com.qc.rc.common.ServerResponse;
 import com.qc.rc.entity.Resume;
 import com.qc.rc.entity.User;
 import com.qc.rc.entity.UserResume;
 import com.qc.rc.entity.pojo.InterviewPojo;
-
+import com.qc.rc.entity.pojo.ResumeInterviews;
 import com.qc.rc.service.InterviewService;
 import com.qc.rc.service.ResumeService;
-
-
-
+import com.sun.org.apache.xml.internal.security.Init;
 
 @Controller
 @RequestMapping("Interview/")
 public class InterviewController {
 	private static Logger logger = Logger.getLogger(InterviewController.class.getName());
-	
-	static Integer userId = 1001;	
 	@Autowired  
 	private InterviewService iInterviewService;
 	
 	@Autowired 
 	private ResumeService resumeService;
+	
+	private static User user = GetUser.getUser();
 	
 	@RequestMapping("selectByCondition.do")
 	public ModelAndView selectByCondition(String pageNum,String interviewInfo,
@@ -57,7 +58,7 @@ public class InterviewController {
 				System.out.println(user.getUserId());
 			}*/
 			Integer pn = 1;
-			Integer st = 0;
+			Integer st = -1;
 			Integer sta = 0;
 			
 			if (StringUtils.isNotBlank(pageNum)) {
@@ -80,7 +81,7 @@ public class InterviewController {
 				sta = Integer.valueOf(status);
 			}
 			
-			PageInfo<InterviewPojo> pageInfo = iInterviewService.selectByCondition(pn, userId, startTime, overTime,
+			PageInfo<InterviewPojo> pageInfo = iInterviewService.selectByCondition(pn, user.getUserId(), startTime, overTime,
 					 interviewJob, interviewInfo, st,sta);
 			model.put("pageInfo", pageInfo);
 			model.put("startTime", startTime);
@@ -90,8 +91,7 @@ public class InterviewController {
 			model.put("sort", st);
 			model.put("status", sta);
 			return new ModelAndView("interviewJsps/showAllInterviews",model);
-			
-			
+
 		} catch (ParseException e) {
 			logger.debug("日期类型不匹配");
 			model.put("message", "日期类型不匹配");
@@ -105,11 +105,10 @@ public class InterviewController {
 			logger.debug("不支持的字符类型转整型");
 			model.put("message", "不支持的字符类型转整型");
 		}
-		return new ModelAndView("interviewJsps/error",model);
-		
-		
-		
+		return new ModelAndView("interviewJsps/error",model);	
 	}
+	
+	
 //	页面跳转
 	@RequestMapping("addNewInterview.do")
 	public ModelAndView addNewInterview(String resumeId,HttpServletRequest request){
@@ -117,8 +116,7 @@ public class InterviewController {
 		logger.debug("a3s2d13as2d");
 		try {
 			if (StringUtils.isNotBlank(resumeId)) {
-				Integer rid = Integer.valueOf(resumeId);
-				Resume resume = resumeService.getResumeDetailsById(rid);
+				Resume resume =  resumeService.getResumeById(resumeId);
 				model.put("resume", resume);
 				return new ModelAndView("interviewJsps/addnewinterview",model);
 	
@@ -147,9 +145,8 @@ public class InterviewController {
 		Map<String,Object> model = new HashMap<String,Object>();
 		try {
 //			临时设置user  应从缓存中得到数据
-			User user =	new User();
-			user.setUserId(1001);
-			user.setUserName("LING");
+			User user =	GetUser.getUser();
+
 
 			String errorMsg = "";
 			if(StringUtils.isBlank(resumeId)){
@@ -190,10 +187,9 @@ public class InterviewController {
 				return new ModelAndView("interviewJsps/error",model);
 			}
 //			从前台获取id查询resume
-			Integer rid = Integer.valueOf(resumeId);
-			Resume resume = resumeService.getResumeDetailsById(rid);
+			Resume resume = resumeService.getResumeById(resumeId);
 
-			Integer result = iInterviewService.addInterview(rid, FormParameterUtil.changeCode(interviewJob), 
+			Integer result = iInterviewService.addInterview(resumeId, FormParameterUtil.changeCode(interviewJob), 
 					FormParameterUtil.changeCode(interviewTime), FormParameterUtil.changeCode(interviewAssociateUsername), FormParameterUtil.changeCode(interviewAssociatePhone),FormParameterUtil.changeCode(interviewAddress) , FormParameterUtil.changeCode(interviewInfo), resume, user);
 			if(result==0){
 				model.put("message","添加面试失败");
@@ -214,7 +210,7 @@ public class InterviewController {
 							logger.debug("发送短信到"+resumePhone);
 							logger.debug(interviewMessage);
 						}
-						return new ModelAndView("redirect:/Interview/selectByCondition.do?userId="+userId+"&sort=1",model);
+						return new ModelAndView("redirect:/Interview/selectByCondition.do",model);
 						//return new ModelAndView("redirect:/Interview/showAllInterviews.do",model);
 					}else {
 						model.put("resume", resume);
@@ -232,7 +228,7 @@ public class InterviewController {
 						logger.debug("发送短信到"+resumePhone);
 						logger.debug(interviewMessage);
 					}
-					return new ModelAndView("redirect:/Interview/selectByCondition.do?userId="+userId,model);
+					return new ModelAndView("redirect:/Interview/selectByCondition.do",model);
 //					转发到详情页面
 //					return new ModelAndView("redirect:/Interview/showInterviewDetail.do?interviewId="+iPojo.getInterviewId(),model);
 				}
@@ -260,9 +256,6 @@ public class InterviewController {
 		Map<String,Object> model = new HashMap<String,Object>();
 		try {
 //			临时设置user  应从缓存中得到数据
-			User user =	new User();
-			user.setUserId(1001);
-			user.setUserName("LING");
 			
 			String errorMsg = "";
 
@@ -305,26 +298,23 @@ public class InterviewController {
 				return new ModelAndView("interviewJsps/error",model);
 			}
 			
-			logger.debug(resumeName+resumePhone+interviewJob+interviewTime+interviewAddress+interviewAssociatePhone+interviewAssociateUsername);
+			
 //			封装数据成resume 在数据库中插入返回resumeid
 			Resume resume = new Resume();
-			
+			resume.setResumeId(GetUuid.getuuid32());
 			resume.setResumeName(FormParameterUtil.changeCode(resumeName));
 			resume.setResumePhone(FormParameterUtil.changeCode(resumePhone));
 			resume.setResumeJobIntension(FormParameterUtil.changeCode(interviewJob));
 			resume.setResumeCreateUser(user.getUserName());
-			System.out.println(resume.toString());
-			int resultCount = resumeService.resumeAdd(resume);
-			System.out.println(resume.toString());
+		
+			int resultCount = resumeService.resumeAddfromInterview(resume);
+		
 //			创建UserResume对象	插入数据库
 			if(resultCount != 0){
-
-				UserResume ur = new UserResume();
-				ur.setUrResumeId(resume.getResumeId());
-				ur.setUrUesrId(userId);
-				int result = resumeService.resumeAddResumeUser(ur);
+				
+				int result = resumeService.resumeAddResumeUser(GetUuid.getuuid32(), user.getUserId(), resume.getResumeId());
 				if(result !=0){
-					System.out.println(resume.toString());
+					
 					result = iInterviewService.addInterview(resume.getResumeId(), FormParameterUtil.changeCode(interviewJob), 
 							FormParameterUtil.changeCode(interviewTime), FormParameterUtil.changeCode(interviewAssociateUsername),
 							FormParameterUtil.changeCode(interviewAssociatePhone)
@@ -346,7 +336,7 @@ public class InterviewController {
 									logger.debug("发送短信到"+resumePhone);
 									logger.debug(interviewMessage);
 								}
-								return new ModelAndView("redirect:/Interview/selectByCondition.do?userId="+userId,model);
+								return new ModelAndView("redirect:/Interview/selectByCondition.do",model);
 								//return new ModelAndView("redirect:/Interview/showAllInterviews.do",model);
 							}else {
 								model.put("resume", resume);
@@ -363,7 +353,7 @@ public class InterviewController {
 								logger.debug("发送短信到"+resumePhone);
 								logger.debug(interviewMessage);
 							}
-							return new ModelAndView("redirect:/Interview/selectByCondition.do?userId="+userId,model);
+							return new ModelAndView("redirect:/Interview/selectByCondition.do",model);
 //							转发到详情页面
 //							return new ModelAndView("redirect:/Interview/showInterviewDetail.do?interviewId="+iPojo.getInterviewId(),model);
 						}
@@ -390,7 +380,10 @@ public class InterviewController {
 	
 	@RequestMapping("deleteById.do")
 	@ResponseBody
-	public Integer deleteById(Integer id){
+	public Integer deleteById(String id){
+		if (StringUtils.isBlank(id)) {
+			return 0;
+		}
 		Integer result = iInterviewService.deleteInterview(id);
 		return result;
 		
@@ -405,5 +398,44 @@ public class InterviewController {
 		ServerResponse<InterviewPojo> interviewPojo = iInterviewService.getInterviewByResumeId(ResumeId);
 		model.put("interviewPojo", interviewPojo.getData());
 		return new ModelAndView("interviewJsps/IVxiangqing",model);	
+	}
+	
+	
+	@RequestMapping("resumeInterviews.do")
+	public ModelAndView getResumeInterviewsByRId(String resumeId){
+		Map<String,Object> model = new HashMap<String,Object>();
+		try {
+			if(StringUtils.isBlank(resumeId)){
+				model.put("message", "resumeId为空");
+				return new ModelAndView("interviewJsps/error",model);
+			}else {
+				ResumeInterviews resumeInterviews = iInterviewService.getResumeInterviewsByRId(resumeId);
+				if(StringUtils.isBlank(resumeInterviews.getResumeName())){
+					model.put("message", "不存在的resumeId");
+					return new ModelAndView("interviewJsps/error",model);
+				}
+				model.put("resumeInterviews", resumeInterviews);
+				return new ModelAndView("interviewJsps/RISdetail",model);	
+			}
+			
+		}catch (NumberFormatException e) {
+			logger.debug("不支持的字符类型转整型");
+			model.put("message", "不支持的字符类型转整型");
+		}
+		return new ModelAndView("interviewJsps/error",model);
+	}
+	
+	
+	@RequestMapping("updateRecodeInfo.do") 
+	@ResponseBody
+	public Integer updateRecodeInfo(String interviewId,String recodeInfo){ 
+		Integer msg = -1;
+		if (StringUtils.isNotBlank(interviewId)) { 
+			System.out.println(interviewId+recodeInfo);
+			msg = iInterviewService.updateInteviewRecodeInfo(recodeInfo, interviewId);
+			
+		}
+		return msg;
+		
 	}
 }
